Fixes to `PreprocessData()` and improvements to user interface.
`PreprocessData()` uses regular expressions to read columns from the input file. These allow some felxibility in the column names, such as "." or space as word separators, and case insensitivity. However, the checks and processing done in the function use the fixed 'legacy' names which means they miss errors if the columns have slighlty different, but valid, names. The update replaces refererences to the fixed legacy names with those found in the actual input file.

Post-processing and results production cannot be done with only one chain. The update allows the user interface to complete as much as possible of the runs. Previously it would fail at the preprocessing stage.

Some warnings from the user interface have been changed to messages. Warnings get burried in the pile of JAGS warnings about unused variables.