Import 'rjags' package wholesale to fix an error finding the `rjags::update.rjags` function.
Set version to 1.5.1 (patch).

Previously, the function call `rjags::update(...)` was used in 'R/runMCMC.R' line 728. This caused the error: "'update' is not an exported object from 'namespace:rjags'", presumably because the function is `update.rjags` (i.e., an S3 method). Importing the whole package and just using the unqualified call `update(...)` is a reversion to commit 514b9967e20baf79df44ef66aac2052d5ecda318 which did not suffer this error.