Storage of run names in argument .RData files changed
When a global run is renamed, the `run_name` element of
'global_mcmc_args.RData', 'post_process_args.RData', and
'combine_runs_args.RData' is edited to be the new run name. Previously, this was
left as the old run name.

The element `rename_list` is now prepended with the original run name only; the
new run name is not in this vector.