Patch to naming of output .csv files when doing median adjustments. Version bumped to 1.5.4.
An error was encountered when saving .csv files when the value of the `run_desc` argument had an underscore at the beginning (e.g., `”_TEST”`). This was due to the auto-shortening of the .csv filenames: double underscores were automatically replaced with single underscores. This was done for the whole path, not just the filename. The result was an invalid path because the directory holding the .csv files is named partly after `run_desc`.